重构受控组件
