+++ lib/createClass.js
@@ -1,5 +1,5 @@
 /**
- * React.createClass补丁 by 司徒正美 Copyright 2018-05-16
+ * React.createClass补丁 by 司徒正美 Copyright 2018-05-17
  * IE9+
  */
