+++ packages/core/createRenderer.js
@@ -1,4 +1,4 @@
-import { extend } from "./util";
+import { extend, noop } from "./util";
 export function createRenderer(methods) {
     return extend(Renderer, methods);
 }
@@ -8,6 +8,7 @@ export const Renderer = {
     mountOrder: 1,
     macrotasks:[],
     boundaries: [],
+    fireDuplex: noop,
     //catchError
     //catchBoundary
     //catchTry
