+++ packages/render/dom/props3.js
@@ -2,11 +2,11 @@ import { NAMESPACE, duplexMap } from "./browser";
 import { patchStyle } from "./style";
 import { eventAction } from "./event";
 import { noop, typeNumber, emptyObject } from "react-core/util";
-import { duplexAction } from "./duplex2";
+import { getDuplexProps, controlledStrategy } from "./duplex";
 
 //布尔属性的值末必为true,false
 //https://github.com/facebook/react/issues/10589
-let rform = /textarea|input|select/i;
+let rform = /textarea|input|select|option/i;
 
 let isSpecialAttr = {
     style: 1,
@@ -120,13 +120,13 @@ function getSVGAttributeName(name) {
 export function diffProps(dom, lastProps, nextProps, fiber) {
     let isSVG = fiber.namespaceURI === NAMESPACE.svg;
     let tag = fiber.type;
-    let continueProps = skipProps;
+    let controlled = "children";
     if (!isSVG && rform.test(fiber.type)) {
-        continueProps = duplexProps;
+        controlled = getDuplexProps(dom, nextProps);
     }
     //eslint-disable-next-line
     for (let name in nextProps) {
-        if (continueProps[name]) {
+        if (name === controlled) {
             continue;
         }
         let val = nextProps[name];
@@ -141,7 +141,7 @@ export function diffProps(dom, lastProps, nextProps, fiber) {
     }
     //如果旧属性在新属性对象不存在，那么移除DOM eslint-disable-next-line
     for (let name in lastProps) {
-        if (continueProps[name]) {
+        if (name === controlled) {
             continue;
         }
         if (!nextProps.hasOwnProperty(name)) {
@@ -153,7 +153,7 @@ export function diffProps(dom, lastProps, nextProps, fiber) {
             actionStrategy[action](dom, name, false, lastProps, fiber);
         }
     }
-    duplexAction(dom, fiber, nextProps, lastProps);
+    controlledStrategy[controlled](dom, controlled, nextProps, lastProps, fiber);
 }
 
 function isBooleanAttr(dom, name) {
@@ -203,19 +203,10 @@ let builtinStringProps = {
     alt: 1,
     lang: 1
 };
-var skipProps = {
-    innerHTML: 1,
-    children: 1,
-}
-var duplexProps = {
-    value: 1,
-    defaultValue: 1,
-    checked: 1,
-    defaultChecked: 1,
-    innerHTML: 1,
-    children: 1,
-}
+
 export let actionStrategy = {
+    innerHTML: noop,
+    children: noop,
     style: function (dom, _, val, lastProps) {
         patchStyle(dom, lastProps.style || emptyObject, val || emptyObject);
     },
@@ -275,7 +266,7 @@ export let actionStrategy = {
                 //如果是假值但不是0，就改成“”,alt不能removeAttribute
                 if (builtinStringProps[name]) {
                     dom[name] = "";
-                } else {
+                }else{
                     dom.removeAttribute(name);
                 }
             } else {
@@ -298,3 +289,4 @@ export let actionStrategy = {
     }
 };
 
+//=============duplex==========
